Some edits regarding the phonics vs. whole-language debate.
These edits are mainly aimed to reflect the fact that the consensus
among psychologists, linguists and neuroscientists is that phonology
plays a crucial role in learning to read.

So, even if whole-language approaches may sometimes work, it's not
because they are consistent with how we think the brain learns to
read, but rather through indirect access to phonological learning.

The full discussion is much more complicated than that, but we're
probably not going to lay it all out here...

This addresses /issues/60.