Checkout.md: specify Etherpad documents
(in fulfilment to the checkout exercise for instructors)
There are several Etherpad documents and two are referred to in this page. This edit aims to forestall confusions.
Section 1 refers to https://pad.carpentries.org/community-discussions and its title is Community Discussions, consistently.
Section 3 refers to https://pad.carpentries.org/teaching-demos and its title is (Carpentries Instructor) Training Demonstration Sessions, inconsistently.